Introduce Webpack
This is a small trial to investigate if I can split out all functionalities and combine them with Webpack.

The problem with the content scripts is that they are not loaded as modules by the browser and therefore it is not possible to require code from other modules. If in the future other platforms besides Strava are to be supported this might be something that may be useful.