Introduce an icon
Introduce an icon as that is mandatory when the extension is offered to Chrome. In honor of Steven Kruijswijks almost victory in the Giro dItalia. He looked so good in that pink jersey it should be mandatory that he wears it the entire year. Just like we should have Ramon Sinkeldam wear the jersey of the Dutch champion.