Prepare Webpack environment
I want to add things like unit tests and linting rules. In order to allow unit testing I want to extract functionalities to separate files. The way user scripts are embedded in a browser I cannot use the Javascript Module functionalities. The solution is to use Webpack to combine the files. To do so the codebase is make NPM ready.

For now nothing changes compared to the previous situation. This is just preparation.

- Introduce `package.json` and `package-lock.json`
- Introduce 'webpack`, `webpack-cli` and `copy-webpack-plugin` as development dependencies
- Introduce scripts in NPM to simplify running webpack - `npm run-script build` - and creating a distribution package - `npm run-script package`
- Introduce `.nvmrc` to indicate Node 14 is the working version for the repository